+++ examples/web/react/provider/Web3Provider.tsx
@@ -22,6 +22,7 @@ export default function Web3ModalProvider({
   return (
     <SophonContextProvider
       network="testnet"
+      authServerUrl="http://localhost:3000"
       partnerId="123b216c-678e-4611-af9a-2d5b7b061258"
       dataScopes={[DataScopes.email]}
     >